requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.1
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.76
onnxruntime-gpu
packaging==23.2
pandas==2.0.3
Pillow==10.0.0
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sympy==1.12
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.158
urllib3==2.0.4
zipp==3.16.2
